+++ changelog.md
@@ -4,6 +4,7 @@
 
 Widgets:
 - Button Widget now supports `options.mustConfirm` to confirm the trigger with custom options.
+- Added support for LX.CalendarRange in DatePicker widget.
 
 Timeline:
 - Clips now have the active property to enable/disable the clip (only visuals).
@@ -14,6 +15,7 @@ Previous Tour masks are now removed. Fixed Tour on resize window.
 Added new `LX.PopConfirm`.
 Fixed PlainText <> symbols in CodeEditor.
 Fixed Calendar `options.fromToday` not including today's date.
+Docs updated.
 
 ## 0.6.11 (master)